Move and group related components to same folder
